Warnings

+++ testing/adios2/engine/staging-common/TestDistributionRead.cpp
@@ -103,7 +103,7 @@ TEST_F(CommonReadTest, ADIOS2CommonRead1D8)
         engine1.Get(step_var, &step);
         engine1.EndStep();
 
-        int result = validateSimpleForwardData(in_R64_1, step, myStart,
+        int result = validateSimpleForwardData(in_R64_1, (int)step, myStart,
                                                myLength, writerSize * Nx);
         if (first_step == SIZE_MAX)
         {
